Created devices/controllers.py and updated devices to use
