+++ tests/devices/unit_tests/test_controllers.py
@@ -1,0 +1,11 @@
+from unittest.mock import Mock
+
+from dodal.devices.controllers import (
+    ConstantDeadTimeController,
+)
+
+
+def test_constant_dead_time_controller_returns_constant():
+    constant_deadtime = 2.4
+    controller = ConstantDeadTimeController(driver=Mock(), deadtime=constant_deadtime)
+    assert controller.get_deadtime(exposure=Mock()) == constant_deadtime
